requirements.txt
# ─── Core Data ───────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1

# ─── Time Series Models ──────────────────────────────────────
statsmodels==0.14.2
pmdarima==2.0.4

# ─── Deep Learning ───────────────────────────────────────────
torch==2.3.0
torchinfo==1.8.0

# ─── MLflow ──────────────────────────────────────────────────
mlflow==2.13.2

# ─── Visualization ───────────────────────────────────────────
matplotlib==3.9.0
seaborn==0.13.2

# ─── Metrics ─────────────────────────────────────────────────
scikit-learn==1.5.0

# ─── Config & CLI ────────────────────────────────────────────
pyyaml==6.0.1
click==8.1.7

# ─── Utilities ───────────────────────────────────────────────
tqdm==4.66.4
joblib==1.4.2
